+++ scripts/tennis2/run_all_expansion.py
@@ -4,22 +4,19 @@ import numpy
 import shutil
 import glob
 
+# ITERATION_NUM=1
 ITERATION_NUM=3
 METHOD_DICT = {
-    'pulse_finetune': {
-        'learning': 'pulse_self_play',
-        'exp_name': 'pulse_finetune',
+    'pulse_expansion_local_op_obs': {
+        'learning': 'pulse_expansion_local_op_obs_self_play_tennis2',
+        'exp_name': 'pulse_expansion_local_op_obs',
         'env': 'env_amp_z',
-        'task': 'HumanoidPP2Z',
+        'task': 'HumanoidTennis2ZLocalOpObs',
     },
 }
 CFG_DIR = "./lma/data/cfg/learning"
 DST_DIR = "./output/HumanoidIm"
-TASK_NAME = "tabletennis2"
-
-INITIAL_WEIGHT_PATH = "./output/HumanoidIm/tabletennis/pulse/Humanoid_00070000.pth"
-RESTART_EPOCH_NUM = 70000
-MAX_EPOCH_NUM = 90000
+TASK_NAME = "tennis2"
 
 
 def main():
@@ -27,14 +24,17 @@ def main():
         for k, v in METHOD_DICT.items():
 
             weight_save_dir = os.path.join(DST_DIR, TASK_NAME, k, str(i))
+
+            if os.path.exists(os.path.join(weight_save_dir, "Humanoid_00020000.pth")):
+                continue
+
             log_save_dir = os.path.join(DST_DIR, TASK_NAME, k, str(i), TASK_NAME, k)
             os.makedirs(weight_save_dir, exist_ok=True)
             os.makedirs(log_save_dir, exist_ok=True)
 
             shutil.copy(os.path.join(CFG_DIR, v["learning"] + ".yaml"), weight_save_dir)
-            shutil.copy(INITIAL_WEIGHT_PATH, weight_save_dir)
 
-            command = ["python", "lma/run_hydra.py"]
+            command = ["python", "phc/run_hydra.py"]
             args = [
                 "project_name=SMPLOlympics",
                 "num_agents=2",
@@ -44,27 +44,24 @@ def main():
                 "env.num_envs=512",
                 "env.task={}".format(v["task"]),
                 "env.enableTaskObs=True",
+                "env.plane.restitution=0.6",
                 "+env.contact_bodies=[\"R_Ankle\",\"L_Ankle\",\"R_Toe\",\"L_Toe\",\"R_Hand\"]",
-                "robot=smpl_humanoid_pp",
+                "robot=smpl_humanoid_tennis_righthand",
                 "robot.has_upright_start=True",
+                "robot.real_weight_porpotion_boxes=False",
+                "env.numAMPObsSteps=10",
                 "env.shape_resampling_interval=500000",
-                "env.motion_file=./sample_data/pingpong1after_upright.pkl",
+                "env.motion_file=./sample_data/video_tennis_afterproc_upright.pkl",
+                "env.episode_length=2000",
                 "headless=True",
                 "env.stateInit=Default",
                 "no_log=True",
-                "epoch={}".format(RESTART_EPOCH_NUM),
-                "learning.params.config.max_epochs={}".format(MAX_EPOCH_NUM)
+                "+learning.params.config.save_initial_weight=true",
             ]
 
             command += args
 
-            print(command)
-
             subprocess.run(command)
-
-            # for p in glob.glob(DST_DIR + '/' + v["exp_name"] + '/*.pth'):
-            #     if os.path.isfile(p):
-            #         shutil.copy(p, weight_save_dir)
 
 if __name__ == '__main__':
     main()
